# Improve IsSavable performance