--- a/Source/Csla/Core/BusinessBase.cs
+++ b/Source/Csla/Core/BusinessBase.cs
@@ -456,6 +456,11 @@ namespace Csla.Core
     {
       get
       {
+        if (!IsDirty || !IsValid || IsBusy)
+        {
+          return false;
+        }
+
         bool auth;
         if (IsDeleted)
           auth = BusinessRules.HasPermission(ApplicationContext, AuthorizationActions.DeleteObject, this);
@@ -463,9 +468,9 @@ namespace Csla.Core
           auth = BusinessRules.HasPermission(ApplicationContext, AuthorizationActions.CreateObject, this);
         else
           auth = BusinessRules.HasPermission(ApplicationContext, AuthorizationActions.EditObject, this);
-        return (auth && IsDirty && IsValid && !IsBusy);
+        return auth;
       }
-    }
+    }    
 
     #endregion