--- a/Source/Csla/Core/BusinessBase.cs
+++ b/Source/Csla/Core/BusinessBase.cs
@@ -456,20 +456,20 @@ namespace Csla.Core
     {
       get
       {
-        if (!IsDirty || !IsValid || IsBusy)
-          return false;
+        var result = IsDirty && IsValid && !IsBusy;
+        if (result)
+        {
+          if (IsDeleted)
+            result = BusinessRules.HasPermission(ApplicationContext, AuthorizationActions.DeleteObject, this);
+          else if (IsNew)
+            result = BusinessRules.HasPermission(ApplicationContext, AuthorizationActions.CreateObject, this);
+          else
+            result = BusinessRules.HasPermission(ApplicationContext, AuthorizationActions.EditObject, this);
+        }
 
-        bool auth;
-        if (IsDeleted)
-          auth = BusinessRules.HasPermission(ApplicationContext, AuthorizationActions.DeleteObject, this);
-        else if (IsNew)
-          auth = BusinessRules.HasPermission(ApplicationContext, AuthorizationActions.CreateObject, this);
-        else
-          auth = BusinessRules.HasPermission(ApplicationContext, AuthorizationActions.EditObject, this);
-        
-        return auth;
+        return result;
       }
-    }    
+    }
 
     #endregion