@rockfordlhotka 2024-05-27
The problem I see is that this could allow someone without permissions to trigger business rules or other checks to run.

Though maybe not, as the saveable check is just looking at pre-set boolean values maybe?