@ossendorf-at-hoelscher 2024-05-28
Yes. The `IsSavable` only looks at bools to determine the state of the object. And if a simple get of a property can trigger rules to run I would be surprised.